# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Official API docs for my Laravel app.",
        "contact": {
            "email": "you@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/payment/process/{order}": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Process payment for a specific order",
                "description": "Initiate the payment process for a specific order using Paymob.",
                "operationId": "processPayment",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID to pay for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount_cents",
                                    "currency",
                                    "delivery_needed",
                                    "shipping_data"
                                ],
                                "properties": {
                                    "amount_cents": {
                                        "type": "string",
                                        "example": "129601148"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "EGP"
                                    },
                                    "delivery_needed": {
                                        "type": "string",
                                        "enum": [
                                            "true",
                                            "false"
                                        ],
                                        "example": "false"
                                    },
                                    "shipping_data": {
                                        "required": [
                                            "first_name",
                                            "last_name",
                                            "phone_number",
                                            "email"
                                        ],
                                        "properties": {
                                            "first_name": {
                                                "type": "string",
                                                "example": "Test"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Account"
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "01010101010"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "test@account.com"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "amount_cents",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "ASC1525"
                                                },
                                                "amount_cents": {
                                                    "type": "string",
                                                    "example": "4000"
                                                },
                                                "quantity": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Smart Watch",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment link generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://accept.paymob.com/payment-request-url"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or order not payable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order Can Not Be Paid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized to pay for this order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized To Pay This Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error from Paymob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed  to place order: Something went wrong"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token to access this endpoint",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}